[spl-tools] [LINUXDAR-135] test failure

+++ spl-tools/jenkins/update.sh
@@ -1,5 +1,7 @@
 #!/bin/bash -e
 
+set -x
+
 #
 # Script to automatically configure or update a Jenkins ubuntu node
 #
@@ -11,16 +13,22 @@ touch /tmp/thisJenkinsJobMadeByViolaGang
 function update_apt()
 {
     echo "=> Updating existing packages with apt..."
-    apt-get update
-    apt-get upgrade -qy
+    apt-get update || exit 1
+    apt-get upgrade -qy || exit 1
 }
 
 function update_yum()
 {
     echo "=> Updating existing packages with yum..."
-    yum -y update
+    yum -y update || exit 1
 }
 
-#( which apt-get &>/dev/null && update_apt ) || update_yum
-( ( which apt-get &>/dev/null && update_yum ) || update_yum ) || rm -rf ./${THIS_DIR} && exit 1
-rm -rf ./${THIS_DIR}
+if [ -n "$(which apt-get)" ]
+then
+    update_yum || rm -rf ./${THIS_DIR} && exit 1
+elif [ -n "$(which yum)" ]
+then
+    update_yum || rm -rf ./${THIS_DIR} && exit 1
+else
+    exit 1
+fi
